Revision of the previous commit based on the notes

+++ test/LargeTests/test_large.py
@@ -1,50 +1,51 @@
 import sys
-sys.path.append('../../')
+from os import path
+sys.path.append(path.dirname(path.dirname(path.dirname(path.abspath(__file__)))))
 from ProjectLib import XMLtoProject
 from TimeLoopLib import TreeDynamicTimeStepping
 import unittest
 import glob
 import os
 
-filepathExampleSetups = "./Test_Setups_large/*.xml"
+filepathExampleSetups = path.join(path.dirname(path.abspath(__file__)),"Test_Setups_large/*.xml")
 xml = glob.glob(filepathExampleSetups)
-e = []
-for xmlfile in xml:
-    resultsFilepath = "./Test_Setup_large/testoutputs/*.*"
-    results = glob.glob(resultsFilepath)
-    for result in results:
-        os.remove(result)
-    print("________________________________________________")
-    print("In the following the setup", xmlfile, "is tested.")
-
-    class MyTest(unittest.TestCase):
-
-        filepathExampleSetups = "./ProjectLib/ExampleSetups/*.xml"
-
-        def test(self):
-            try:
-                prj = XMLtoProject(xml_project_file=xmlfile)
-                time_stepper = TreeDynamicTimeStepping(prj)
-                prj.runProject(time_stepper)
-            except:
-                self.fail(e.append(xmlfile))
-
-    if __name__ == "__main__":
-        unittest.main(exit=False)
-    print("The setup", xmlfile, "was tested.")
-    print("________________________________________________")
-print("The testing of all example setups is finished")
-print("")
-if e:
-    if len(e) == 1:
-        print("An error occured while testing the following example setup:")
-    else:
-        print("Errors occured while testing the following example setups:")
-    n = range(len(e))
-    for x in n:
+errors = []
+if xml:
+    for xmlfile in xml:
+        resultsFilepath = path.join(path.dirname(path.abspath(__file__)),"Test_Setups_large/testoutputs/*.*")
+        results = glob.glob(resultsFilepath)
+        for result in results:
+            os.remove(result)
+        print("________________________________________________")
+        print("In the following the setup", xmlfile, "is tested.")
+    
+        class MyTest(unittest.TestCase):
+    
+            def test(self):
+                try:
+                    prj = XMLtoProject(xml_project_file=xmlfile)
+                    time_stepper = TreeDynamicTimeStepping(prj)
+                    prj.runProject(time_stepper)
+                except:
+                    self.fail(errors.append(xmlfile))
+    
+        if __name__ == "__main__":
+            unittest.main(exit=False)
+        print("The setup", xmlfile, "was tested.")
+        print("________________________________________________")
+    print("The testing of all example setups is finished")
+    print("")
+    if errors:
+        if len(errors) == 1:
+            print("An error occured while testing the following example setup:")
+        else:
+            print("Errors occured while testing the following example setups:")
+        n = range(len(errors))
+        for x in n:
+            print("")
+            print(errors[x])
         print("")
-        print(e[x])
-    print("")
-else:
-    print("The tests of all example setups were successful")
-    print("")
+    else:
+        print("The tests of all example setups were successful")
+        print("")
+else: print("unfortunately no project-file could be found.")
